requirements.txt
requests==2.31.0
urllib3==1.25.7
botocore==1.34.58
s3transfer==0.10.0
boto3==1.34.58
awscli==1.32.58
pandas==2.0.3
toml==0.10.2
python-dotenv==1.0.1
SQLAlchemy==2.0.28
mysql-connector-python==8.3.0